chore: proposal

+++ src/proposal.ts
@@ -18,10 +18,11 @@ import { Auth } from "."
  * @class Proposal
  * @extends Client
  */
-export class Proposal extends Client {
+export class Proposal {
   private static _config: Maybe<{ devMode: boolean }> = null
 
   private static _instance: Maybe<Proposal> = null
+  private static _client: Maybe<Client>
 
   /**
    * Get the PROPOSAL_STATUS constant object.
@@ -39,13 +40,6 @@ export class Proposal extends Client {
     return { ...PROPOSAL_TYPE }
   }
 
-  private constructor() {
-    if (!!!Proposal._config)
-      throw new Error("Proposal must be configured before getting the instance")
-
-    super(Proposal._config.devMode)
-  }
-
   /** static methods */
 
   static config(config: { devMode: boolean }) {
@@ -58,7 +52,12 @@ export class Proposal extends Client {
     return Proposal._instance ?? new Proposal()
   }
 
-  /** private instance methods */
+  private constructor() {
+    if (!!!Proposal._config)
+      throw new Error("Proposal must be configured before getting the instance")
+
+    Proposal._client = new Client(Proposal._config.devMode)
+  }
 
   /**
    * Creates a new proposal object and inserts it into the backend.
@@ -67,14 +66,16 @@ export class Proposal extends Client {
    * @throws {Error} If signedMessage is required but not provided.
    */
   private async _createProposal(proposal: CreateProposal): Promise<boolean> {
+    if (!!!Proposal._config || !!!Proposal._instance || !!!Proposal._client)
+      throw new Error("Proposal has not been configured")
+
     try {
-      const { response, statusCode } = await this._fetch<ApiResponse<boolean>>(
-        this._backendUrl("/proposal/insert"),
-        {
-          method: "POST",
-          body: proposal,
-        }
-      )
+      const { response, statusCode } = await Proposal._client.fetch<
+        ApiResponse<boolean>
+      >(Proposal._client.backendUrl("/proposal/insert"), {
+        method: "POST",
+        body: proposal,
+      })
 
       if (statusCode !== 200 || !response || !response.data) return false
 
@@ -100,11 +101,13 @@ export class Proposal extends Client {
    * @throws {Error} If the "id" parameter is invalid or if an error occurs during the retrieval process.
    */
   async get(id: string, did?: string): Promise<Maybe<IProposal>> {
+    if (!!!Proposal._config || !!!Proposal._instance || !!!Proposal._client)
+      throw new Error("Proposal has not been configured")
     if (!id) throw new Error('Invalid parameter "id".')
 
     try {
-      const { response } = await this._fetch<ApiResponse<IProposal>>(
-        this._backendUrl(`/proposal/${id}${did ? `/${did}` : ``}`)
+      const { response } = await Proposal._client.fetch<ApiResponse<IProposal>>(
+        Proposal._client.backendUrl(`/proposal/${id}${did ? `/${did}` : ``}`)
       )
 
       if (!response || !response.data) return null
@@ -139,6 +142,9 @@ export class Proposal extends Client {
     take?: number,
     did?: string
   ): Promise<Maybe<ListProposalsResponse>> {
+    if (!!!Proposal._config || !!!Proposal._instance || !!!Proposal._client)
+      throw new Error("Proposal has not been configured")
+
     const filtersInput = filtersOptions ? { ...filtersOptions } : null
 
     let filters: any = null
@@ -190,10 +196,10 @@ export class Proposal extends Client {
     }
 
     try {
-      const { response } = await this._fetch<
+      const { response } = await Proposal._client.fetch<
         ApiResponse<ListProposalsResponse>
       >(
-        this._backendUrl(
+        Proposal._client.backendUrl(
           `/proposals/${skipUrl}/${takeUrl}${did ? `/${did}` : ``}`
         ),
         {
@@ -235,13 +241,19 @@ export class Proposal extends Client {
    * @throws {Error} If the signedMessage is required but not provided.
    */
   async delete(id: string, creatorAddress: string): Promise<void> {
+    if (!!!Proposal._config || !!!Proposal._instance || !!!Proposal._client)
+      throw new Error("Proposal has not been configured")
+
     try {
-      await this._fetch(this._backendUrl(`/proposal/${id}/delete`), {
-        method: "DELETE",
-        body: {
-          creatorAddress,
-        },
-      })
+      await Proposal._client.fetch(
+        Proposal._client.backendUrl(`/proposal/${id}/delete`),
+        {
+          method: "DELETE",
+          body: {
+            creatorAddress,
+          },
+        }
+      )
     } catch (error: any) {
       console.warn(error)
